# docs/API/openapi.yaml
openapi: 3.1.0
info:
  title: Genesis 360 Empresarial API
  version: 1.2.1-rc.1
  description: "Contratos HTTP da baseline V1. Autorização de tenant é validada no servidor; IDs enviados pelo cliente\
    \ nunca substituem o tenant ativo autenticado.\n V1.2.1: mutações por cookie exigem Origin igual ao host servido;\
    \ papéis autorizados por operação; corpo JSON limitado a 64 KiB. Material sensível é somente declarado pela\
    \ entrada pública."
servers:
- url: /api
...
paths:
...
  /consents:
    post:
      tags:
      - Consent
      operationId: recordConsentDecision
      summary: Registra decisão append-only por finalidade e versão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConsentDecisionInput'
      responses:
        '200':
          description: Decisão registrada
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
...
  /evidence:
    get:
      tags:
      - Evidence
      operationId: listEvidence
      parameters:
      - name: companyId
        in: query
        required: true
        schema:
          type: string
          format: uuid
      responses:
        '200':
          description: Evidências autorizadas
    post:
      tags:
      - Evidence
      operationId: recordEvidence
      summary: Registra evidência não verificada e vínculo opcional
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EvidenceInput'
      responses:
        '201':
          description: Evidência registrada
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
  /diagnostics:
    post:
      tags:
      - Diagnostic
      operationId: startDiagnostic
      summary: Inicia diagnóstico usando versão canônica publicada
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
              - companyId
              properties:
                companyId:
                  type: string
                  format: uuid
                profile:
                  type: string
                  enum:
                  - ESSENTIAL
                  - FULL
                  default: ESSENTIAL
      responses:
        '201':
          description: Diagnóstico iniciado
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
...
  /decisions/{id}/missions:
    post:
      tags:
      - Mission
      operationId: createMissionFromDecision
      summary: Cria uma missão versionada a partir do GDS
      parameters:
      - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              type: object
              properties:
                dueAt:
                  anyOf:
                  - type: string
                    format: date-time
                  - type: 'null'
      responses:
        '201':
          description: Missão criada
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
...
  /missions/{id}/evidence:
    post:
      tags:
      - Mission
      operationId: recordMissionEvidence
      summary: Registra evidência da missão e alimenta Evidence Ledger
      parameters:
      - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionEvidenceInput'
      responses:
        '201':
          description: Evidência registrada
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
  /missions/{id}/outcome:
    post:
      tags:
      - Mission
      operationId: recordMissionOutcome
      summary: Registra outcome e fecha Mission → Timeline
      parameters:
      - $ref: '#/components/parameters/ResourceId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MissionOutcomeInput'
      responses:
        '201':
          description: Outcome registrado
        '409':
          description: Conflito de estado/revisão; reler antes de tentar gravação.
        '403':
          description: Origem, tenant ou papel não autorizados; nunca expõe detalhes internos.
        '413':
          description: Corpo acima de 64 KiB.
        '500':
          description: Falha interna sem vazamento de mensagens do provedor.
...
components:
  parameters:
    ResourceId:
      name: id
      in: path
      required: true
      schema:
        type: string
        format: uuid
  responses:
    BadRequest:
      description: Entrada inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Sessão inválida
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
  schemas:
    ErrorResponse:
      type: object
      required:
      - error
      properties:
        error:
          type: string
    ConsentDecisionInput:
      type: object
      required:
      - purposeCode
      - granted
      - consentVersionId
      properties:
        purposeCode:
          type: string
          enum:
          - CORE_OPERATION
          - AI_PROCESSING
          - QUALIFIED_MATCHING
          - ECOSYSTEM_ANALYTICS
          - BENCHMARK_ANALYTICS
        granted:
          type: boolean
        consentVersionId:
          type: string
          format: uuid
        companyId:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
...
    EvidenceInput:
      type: object
      required:
      - companyId
      - evidenceType
      - summary
      properties:
        companyId:
          type: string
          format: uuid
        evidenceType:
          type: string
          enum:
          - user_declaration
          - metric
          - observation
        summary:
          type: string
        payload:
          type: object
        sourceRef:
          anyOf:
          - type: string
          - type: 'null'
        confidence:
          type: 'null'
        sensitivity:
          type: string
          enum:
          - public
          - internal
          - personal
          - financial
          - restricted
        purposeCodes:
          type: array
          items:
            type: string
      additionalProperties: false
...
    MissionEvidenceInput:
      type: object
      required:
      - evidenceType
      - summary
      properties:
        evidenceType:
          type: string
          enum:
          - declaration
          - metric
          - link
        summary:
          type: string
        payload:
          type: object
        sourceRef:
          anyOf:
          - type: string
          - type: 'null'
    MissionOutcomeInput:
      type: object
      required:
      - outcomeStatus
      properties:
        outcomeStatus:
          type: string
          enum:
          - no_change
          - small_improvement
          - improved
          - significant_improvement
          - worsened
          - unknown
        beforeValue: {}
        afterValue: {}
        metricCode:
          anyOf:
          - type: string
          - type: 'null'
        source:
          type: string
          enum:
          - declared
          default: declared
        confidence:
          type: 'null'
          description: Confiança não pode ser autodeclarada.
      additionalProperties: false
...